CROSSEVAL-33 Assignment API
Created API for assignment create/update. If it gets assignment_id then you can give additional data and change the assignment. And if not It creates new assignment. Also after creating new assignment it also adds all users of this course for the assignment. So every student will have this assignment in his page.