CROSSEVAL-38 API for supervisor grade
Created API for supervisor side grade showing. It will show the grades for the all students that are registered to his course.

Also changed the parameters for the users fields, so there is need only few(password, redem_password, isSupervisor and email) fields. Others will be added in profile page. Which API will be done in different API.