CROSSEVAL-61 missed assignments API
Added command which will run every minute and check if there are the assignments that are ended and needs to check the students to submit their work. If the student didn't submit the work, then the command will change their status to “missed”.
Also, I added the API to check the same logic, but the difference between them is only in the way how they find the assignments. The API type code checks all assignments that are past the current time, but the command would check the assignments that were past the hour, from their end.