CROSSEVAL-62 Excel grades export
Added Excel formatting framework for Laravel.

Added Export code for the formatting and filling the table with appropriate data which is passed to BackEnd by API.

Added styles to the Excel file. So, the column’s size will automatically adjust to the text size, and headings will be in bold style.

Changed, and added for Assignments and Answers table relations with their foreign keys.

In the end, the file name would be something like that: “grades_MDE338“.

Important Notes:
The parameters for this GET API should be in this format:

”course_id” = 1
“assignment_ids“ = “1,2,3,4,5,6“

So, in the end, the API request should look like this:
http://127.0.0.1:8000/api/v1/grades/export?course_id=1&assignment_ids=2,3,4,5