CROSSEVAL-71 Students list API
Created Resource which will help to show every student that registered into course. And will also work for just Users too. And the API itself will also work in the way where it also checks if the user trying to access the API is the supervisor of the requesting course.